pipeline/pyproject.toml
[project]
name = "woograph"
version = "0.1.0"
description = "Collaborative knowledge graph builder"
requires-python = ">=3.11"
dependencies = [
    "pymupdf4llm>=0.0.10",
    "trafilatura>=1.12",
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "click>=8.1",
    "spacy>=3.7",
    "anthropic>=0.40",
]

[project.scripts]
woograph = "woograph.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/woograph"]

[tool.ruff]
target-version = "py311"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4"
]
